requirements.txt
pandas>=2.0,<3.0
numpy>=1.24,<3.0
sqlalchemy>=2.0,<3.0
scikit-learn>=1.3,<2.0
PyYAML>=6.0,<7.0
joblib>=1.3,<2.0
pytest>=8.0,<9.0
streamlit>=1.35,<2.0
